Move `sources.py` to examples.
